feat: full zod type inference
