+++ packages/core/src/provider/schema.ts
@@ -1,0 +1,62 @@
+import { Address as zAddress } from "abitype/zod";
+import type { Chain, Transport } from "viem";
+import z from "zod";
+import type { PublicErc4337Client, SupportedTransports } from "../client/types";
+import { getChain } from "../utils/index.js";
+
+export const SmartAccountProviderOptsSchema = z.object({
+  /**
+   * The maximum number of times to try fetching a transaction receipt before giving up (default: 5)
+   */
+  txMaxRetries: z.number().optional().default(5),
+
+  /**
+   * The interval in milliseconds to wait between retries while waiting for tx receipts (default: 2_000)
+   */
+  txRetryIntervalMs: z.number().optional().default(2_000),
+
+  /**
+   * The mulitplier on interval length to wait between retries while waiting for tx receipts (default: 1.5)
+   */
+  txRetryMulitplier: z.number().optional().default(1.5),
+
+  /**
+   * used when computing the fees for a user operation (default: 100_000_000n)
+   */
+  minPriorityFeePerBid: z.bigint().optional().default(100_000_000n),
+});
+
+export const SmartAccountProviderConfigSchema = <
+  TTransport extends SupportedTransports = Transport
+>() =>
+  z.object({
+    rpcProvider: z.union([
+      z.string(),
+      z
+        .any()
+        .refine<PublicErc4337Client<TTransport>>(
+          (provider): provider is PublicErc4337Client<TTransport> => {
+            return (
+              typeof provider === "object" &&
+              "request" in provider &&
+              "type" in provider &&
+              "key" in provider &&
+              "name" in provider
+            );
+          }
+        ),
+    ]),
+    chain: z.any().refine<Chain>((chain): chain is Chain => {
+      if (!("id" in chain) || typeof chain.id !== "number") {
+        return false;
+      }
+
+      try {
+        return getChain(chain.id) !== undefined;
+      } catch {
+        return false;
+      }
+    }),
+    entryPointAddress: zAddress,
+    opts: SmartAccountProviderOptsSchema.optional(),
+  });
